requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
aiofiles>=23.0.0
websockets>=12.0
requests>=2.28.0
duckduckgo-search>=4.0.0
bing-image-downloader>=1.1.0
